+++ WAC Aircraft/lua/entities/wac_hc_base/cl_init.lua
@@ -1,6 +1,5 @@
 
 
-
 include("shared.lua")
 
 
@@ -556,7 +555,6 @@ function ENT:Draw()
 					particle:SetEndSize(20*self.Scale)
 					particle:SetColor(255,255,255)
 					particle:SetRoll(math.Rand(-50,50))
-					//self.Emitter:Finish()
 				end
 			else
 				local particle = self.Emitter:Add("sprites/heatwave",self:LocalToWorld(self.SmokePos))
@@ -568,7 +566,6 @@ function ENT:Draw()
 				particle:SetEndSize(20*self.Scale)
 				particle:SetColor(255,255,255)
 				particle:SetRoll(math.Rand(-50,50))
-				//self.Emitter:Finish()
 			end
 			self.lastHeatDrawn = CurTime()
 		end
@@ -591,4 +588,5 @@ net.Receive("wac.aircraft.updateWeapons", function(length)
 		end
 	end
 end)
+