+++ WAC Aircraft/lua/entities/wac_hc_hebullet/init.lua
@@ -4,7 +4,7 @@ AddCSLuaFile("shared.lua")
 include("shared.lua")
 
 function ENT:Initialize()
-	self.Entity:SetModel("models/props_junk/PopCan01a.mdl")		
+	self.Entity:SetModel("models/props_junk/popcan01a.mdl")		
 	self.Entity:PhysicsInit(SOLID_VPHYSICS)
 	self.Entity:SetMoveType(MOVETYPE_VPHYSICS)
 	self.Entity:SetSolid(SOLID_VPHYSICS)