+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Version 338
+#### Aircraft
+- Added maintenance station
+
 ## Version 337
 #### Aircraft
 - Fixed helicopter thrust being too high at lower tickrate